pyBackend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
requests>=2.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0
aiofiles>=23.0.0
python-multipart>=0.0.5
pytz>=2023.3
